# use bash -c in hc run_cmd