--- a/crates/cli/src/config_files/build.rs
+++ b/crates/cli/src/config_files/build.rs
@@ -34,7 +34,7 @@ impl Build {
         for build_step in &self.steps {
             let slice_vec: Vec<_> = build_step.arguments.iter().map(|e| e.as_str()).collect();
             util::run_cmd(
-                base_path.to_path_buf(),
+                &base_path.to_path_buf(),
                 build_step.command.clone(),
                 &slice_vec[..],
             )?;