--- a/crates/cli/src/util.rs
+++ b/crates/cli/src/util.rs
@@ -40,7 +40,7 @@ pub fn run_cmd(base_path: &PathBuf, bin: String, args: &[&str]) -> DefaultResult
     //  "command": "wasm-gc",
     //  "arguments": ["$CARGO_TARGET_DIR/wasm32-unknown-unknown/release/{{ name }}.wasm"]
     // },
-    // note the explicit (in cargo) and implicit (in wasm-gc) use of $CARGO_TARGET_DIR!
+    // note the implicit (in cargo) and explicit (in wasm-gc) use of $CARGO_TARGET_DIR!
     // note also that the supported development environment is based on `nix-shell` so we at least
     // assume/expect that developers are using something bash-like for development already
     // e.g. we assume `cargo`, `wasm-gc`, `wasm-opt`, `wasm2wat`, `wat2wasm` all exist in the