## PR summary

## testing/benchmarking notes

( if any manual testing or benchmarking was/should be done, add notes and/or screenshots here )

manual testing:

before, passing `"$PWD"` to `run_cmd`:

```
[nix-shell:~/holochain-rust]$ cargo test -p hc run_cmd_test -- --nocapture
   Compiling hc v0.0.41-alpha4 (/home/thedavidmeister/holochain-rust/crates/cli)
    Finished test [unoptimized + debuginfo] target(s) in 7.07s
     Running target/debug/deps/hc-a4319c0a8f06a72c

running 1 test
> echo $PWD
"echo $PWD"
$PWD
test util::tests::run_cmd_test ... ok

test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 10 filtered out
```

before, passing `"\"$PWD\""` to `run_cmd`:

```
[nix-shell:~/holochain-rust]$ cargo test -p hc run_cmd_test -- --nocapture
   Compiling hc v0.0.41-alpha4 (/home/thedavidmeister/holochain-rust/crates/cli)
    Finished test [unoptimized + debuginfo] target(s) in 6.76s
     Running target/debug/deps/hc-a4319c0a8f06a72c

running 1 test
> echo "$PWD"
"echo \"$PWD\""
"$PWD"
test util::tests::run_cmd_test ... ok

test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 10 filtered out
```

after, passing `"$PWD"` to `run_cmd`:

```
[nix-shell:~/holochain-rust]$ cargo test -p hc run_cmd_test -- --nocapture
   Compiling hc v0.0.41-alpha4 (/home/thedavidmeister/holochain-rust/crates/cli)
    Finished test [unoptimized + debuginfo] target(s) in 6.91s
     Running target/debug/deps/hc-a4319c0a8f06a72c

running 1 test
> echo $PWD
"echo $PWD"
/tmp/tmp.AVS17kzpYd/org_holochain_testaHkIaA
test util::tests::run_cmd_test ... ok

test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 10 filtered out
```

after, passing `"\"$PWD\""` to `run_cmd`:

```
[nix-shell:~/holochain-rust]$ cargo test -p hc run_cmd_test -- --nocapture
   Compiling hc v0.0.41-alpha4 (/home/thedavidmeister/holochain-rust/crates/cli)
    Finished test [unoptimized + debuginfo] target(s) in 6.86s
     Running target/debug/deps/hc-a4319c0a8f06a72c

running 1 test
> echo "$PWD"
"echo \"$PWD\""
/tmp/tmp.AVS17kzpYd/org_holochain_testemzqDO
test util::tests::run_cmd_test ... ok

test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 10 filtered out
```

after, passing `"'$PWD'"` (single quotes) to `run_cmd`:

```
[nix-shell:~/holochain-rust]$ cargo test -p hc run_cmd_test -- --nocapture
   Compiling hc v0.0.41-alpha4 (/home/thedavidmeister/holochain-rust/crates/cli)
    Finished test [unoptimized + debuginfo] target(s) in 6.77s
     Running target/debug/deps/hc-a4319c0a8f06a72c

running 1 test
> echo '$PWD'
"echo \'$PWD\'"
$PWD
test util::tests::run_cmd_test ... ok

test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 10 filtered out
```

## followups

( any new tickets/concerns that were discovered or created during this work but aren't in scope for review here )

## changelog

- [ ] if this is a code change that effects some consumer (e.g. zome developers) of holochain core,  then it has been added to [our between-release changelog](https://github.com/holochain/holochain-rust/blob/develop/CHANGELOG-UNRELEASED.md) with the format 

```markdown
- summary of change [PR#1234](https://github.com/holochain/holochain-rust/pull/1234)
```

## documentation

- [ ] this code has been documented according to our [docs checklist](https://hackmd.io/@freesig/Hk9AmKJNS)